+++ changelog.md
@@ -1,12 +1,14 @@
 ___
 ## v2.2
 ```
-DATE: 2021_04.2
+DATE: 2021_05
 ```
 
 + Change the format GO categories based on the evidences codes.
 
 + Rewrite the obsolete code in the calculation of cross-references.
+
++ Remote the version in the identifiers (RefSeq and GENCODE)
 
 ___
 ## v2.1